refactor: move connection/channel state types out of protocol.rs
protocol.rs mixed the public realtime state model with the internal wire
protocol. Move the public types to their owning domain modules and leave
protocol.rs wire-only:

- ConnectionState/ConnectionEvent/ConnectionStateChange -> connection/mod.rs
- ChannelState/ChannelEvent/ChannelStateChange/ChannelMode -> channel.rs

protocol.rs now holds only ProtocolMessage and its supporting wire types, and
is demoted to `pub(crate) mod` (it exposed no public items of its own). The
crate-root re-exports are unchanged, so the public API (`ably::ConnectionState`,
etc.) is byte-for-byte identical; only internal paths moved.

Import sites now take the state types via the crate-root re-export and keep
wire types on `crate::protocol::`.
